DAC ramp demo: Change buffer length calculation
Splits the shift left by two into two multiplies on either side of the float-to-int conversion. This means the length of one ramp for-loop is calculated then multiplied by two.

This solves a bug where some values of amplitude and step could cause buffer overflows, and was found through user code with amplitude 5 and step 2 (calculating a buffer length of 8 instead of 10), here: https://forum.digilentinc.com/topic/22675-zmod-adc-and-dac-initialization/

The overflow led to some pointer being overwritten prior to being freed, which caused the heap to underflow, and all further malloc calls to fail.